Implmented a bitwise loop solution
This generates a result by doing a bitwise or of the two inputs,
Then using the bitwise and we can see what bits were counted twice.

Subtracting this set of bits that was counted twice and then adding it back
on shifted one bit over lets us get closer to the real solution.
Each time we do this subtract, shift, add we update the remainder.

looping until the remainder is 0, worst case this will loop 31 times.
e.g. int.MaxValue + 1